sync

+++ src/xdpd/drivers/gnu_linux_dpdk_ng/src/io/iface_manager.cc
@@ -1336,7 +1336,7 @@ rofl_result_t iface_manager_discover_physical_ports(void){
 		// configure transmit queues
 		uint16_t nb_tx_desc = dev_info.tx_desc_lim.nb_max / tx_lcores[socket_id].size();
 		if (not phyports[port_id].is_virtual && iface_manager_port_setting_exists(s_pci_addr, "nb_tx_desc")) {
-			nb_tx_desc = iface_manager_get_port_setting_as<uint64_t>(s_pci_addr, "nb_tx_desc");
+			nb_tx_desc = iface_manager_get_port_setting_as<uint16_t>(s_pci_addr, "nb_tx_desc");
 		}
 
 		uint64_t tx_prefetch_threshold(TX_PREFETCH_THRESHOLD_DEFAULT);
@@ -1436,7 +1436,7 @@ rofl_result_t iface_manager_discover_physical_ports(void){
 		// configure receive queues
 		uint16_t nb_rx_desc = dev_info.rx_desc_lim.nb_max / rx_lcores[socket_id].size();
 		if (not phyports[port_id].is_virtual && iface_manager_port_setting_exists(s_pci_addr, "nb_rx_desc")) {
-			nb_rx_desc = iface_manager_get_port_setting_as<uint64_t>(s_pci_addr, "nb_rx_desc");
+			nb_rx_desc = iface_manager_get_port_setting_as<uint16_t>(s_pci_addr, "nb_rx_desc");
 		}
 
 		uint64_t rx_prefetch_threshold(RX_PREFETCH_THRESHOLD_DEFAULT);
